fix bug
CGContextSetStrokeColorWithColor: invalid context 0x0. This is a
serious error. This application, or a library it uses, is using an
invalid context  and is thereby contributing to an overall degradation
of system stability and reliability. This notice is a courtesy: please
fix this problem. It will become a fatal error in an upcoming update.